+++ BackEnd/server/src/main/java/com/NePool/app/service/CommentService.java
@@ -13,7 +13,7 @@ public interface CommentService {
     CommentRequestDTO register(String username, String work_book_id, CommentRequestDTO dto) throws Exception;
 
     PageResultDTO<CommentRequestDTO,Comments> getList(String work_book_id, PageRequestDTO dto) throws Exception;
-    double getLike(String work_book_id) throws Exception;
+    Float getLike(String work_book_id) throws Exception;
 
     String delete(String comment_id,String writer) throws Exception;
     default Comments dtoToEntity(CommentRequestDTO dto, String workBook_id, String user_id,String id) {
